frontend/requirements.txt
pandas==2.0.3
PyYAML==6.0.1
Requests==2.31.0
streamlit==1.27.2
joblib==1.3.2
optuna==3.4.0
seaborn==0.13.0
matplotlib==3.8.0
plotly==5.18.0
scikit_learn==1.3.1
